+++ application/src/main/java/org/opentripplanner/apis/gtfs/support/time/LocalDateRangeUtil.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.apis.gtfs.support.time;
 
 import org.opentripplanner.apis.gtfs.generated.GraphQLTypes;
-import org.opentripplanner.apis.gtfs.model.LocalDateRange;
+import org.opentripplanner.core.model.time.LocalDateRange;
 
 public class LocalDateRangeUtil {
