+++ domain-core/src/test/java/org/opentripplanner/core/model/time/LocalDateRangeTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.apis.gtfs.model;
+package org.opentripplanner.core.model.time;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
